Implementation of Day Schedulers (beta). An addition to choosing order and amount of scheduling.
Day Schedulers implement the basics of scheduling used in Operating
Systems. The algorithms FirstComeFirstServe, ShortestJobFirst (or task
in this case), and RoundRobin are implemented.

FirstComeFirstServe is the standard algorithm that was already in use.
Events are schedule in the order they are added to the collection that
stores the events for scheduling.

ShortestJobFirst simply orders the events by the amount of time they
require of scheduling.

RoundRobin uses the concept of a quantum to not allow an event to span
continuously over an amount of time or quantum. Events are alternated
for the amount of quantum specified.

ChangeLog:
-------------------

+ Added package 'scheduling.dayScheduling'
| • Contains all DayScheduler base and structure.
| - Must optimize code for expansion.

+ Added SchedulerTester abstract class.
| • Provides a basis for developing tests on a scheduler.
| • As JavaDoc describes, requires only a few methods to implement.
| • Automatic connection of methods done in abstraction.

+ Added 'test.scheduling' package.
| • TestCase01 modified for SchedulerTester abstract class.
| • TestCase02 tests the ShortestTaskFirst algorithm.
| • TestCase03 tests the RoundRobin algorithm.